Renderer: Allow slicing the viewport
Introduce a new type called "RendererTargetView".
This type can temporarily take ownership of a target, restrict and map the drawing surface and
later release the original target for further use.

This allows you to create "offsets" for other objects with ease by creating a view at the desired offset,
drawing the shape there and releasing the original RendererTarget again.