dns: Support A and CNAME records
Basic DNS resolution works now. The code is a bit messy,
especially the variable names need some cleanup.
Still, its good enough to resolve a given domain to an IP Address.

Exciting ^^

Cleanup will be done in a later commit.

Caching is not yet supported.

If domain resolution requires resolving
another domain that is not in the current answer, we fail.