dns: fix two bugs in punycode implementation
We were inserting decoded characters at a specific byte position.
However, in unicode, byte boundaries must not necessary also be
character boundaries. This was fixed by using a Vec<char> and later
converting that to a String.

The adapt() function was also incorrect. It contained a multiplication
by BASE that was not necessary.

These bugs were found with the punycode spec tests that will be added in
a later commit.