dns: Use punycode in dns lookups and recurse through nameservers
We now automatically use punycode encoding behind the scences to allow
lookup of non-ascii domains.

We also use the "authority" and "additional" sections of the answer
to properly lookup a domain.

Before, we used a NS and just assumed it was authoritative for every
single domain (which obviously isn't the case)
Now, a global root server is used to lookup the domain
segment-by-segment.