html: work around the lack of inheritance for DOM
This commit introduces a way to imitate inheritance in the way it is
required for IDL interfaces in various specs. It is still very raw and
not really safe but i hope to use a lot more macros in the future so
developers have less chances to screw things up.

Every IDL interface contains its superclass as a first field with
\#[repr(C)]. That means that pointers can safely be casted along the
class hierarchy. But, it relies on the developer accurately
defining the class hierarchy (to ensure safe downcasting)

All in all this is just a rough draft, expect lots of changes in the
future.

This approach is similar to what serde does. I will probably need to
adapt it in the future as the DOM needs to be managed by the JS runtime
but thats a problem for another day.