Gui: Replace previous backend system with SDL2 Adapter
We are scaling the project down a bit. Writing a full on wayland client would take months
and even then, we would *still* only be running on wayland!

Instead, we are using SDL2. The compromise is that we are wrapping all the SDL2 stuff in a way
that can easily be extended to other backends later.

For now, all SDL2 features are enabled because otherwise compilation fails, this might be a bug
and should be investigated later (we don't need features like "ttf" or "image")