gui: Allow statically sized objects
Previously, all sizes had to be relative to the window size, ie. 70/30 or something like that.

Now, elements can specify that they want to be exactly X pixels large, no matter the size of the window
and the Divider takes care of that.

The divider also supports an arbitrary number of children now.