gui: remove unnecessary abstraction over SDL2
I originally intended to use my own traits and gui types (like rect, color, Surface etc)

This decision was made so the engine could easily be ported to a different backend.

However, the more i work on it, the more of a pain it becomes to write an abstraction for
every type thats ultimately just a very thin translation of the SDL type.

SDL is fine. If we want to support something else later, we can worry about it later too.