base: remove custom log crate
Writing our own logger crate is just silly. While i enjoy building things from scratch,
i wasn't really using the logger anyways and it was a pain to work with.

I am building a browser. If i want to get to a MVP state this century, i need to keep
the scope manageable.