hash: use SIMD instructions when available
This provides a ~20x speedup over the naive implementation.
I could also do this for ARM since they apparently have CRC32 intrinsics but
I don't have an ARM machine so that will have to be done later.

This link first inspired me to implement this change:
https://nigeltao.github.io/blog/2021/fastest-safest-png-decoder.html#crc-32

To be fair, at this point, we really don't care about the runtime of crc32.
But it was fun to implement anyways :)