font + image + gui: Refactor crate hierarchy to allow for a generic Canvas type
To prevent dependency cycles, i outsourced a pixel-access canvas type to a new "canvas" crate.
This canvas is imported and written to by "font" and "image" and read by "gui".

This allows us to render fonts with the new sdl backend!

Also some font render changes which i can't be bothered to put into a seperate commit.