font: Adapt rasterizer to the new "Font" type.
Also prettified/fixed a lot of the ugly font code. I wrote that
before really understanding the ttf format so there
were (and are still) a lot of easy improvements to be made.

(Like the Loca table being only used by the Glyf table and the Format4 table not
exposing all of its private methods since they are not needed anywhere else)