gui + browser: rethink the way applications work
Previously, we would create an SDL2 Window and some layout types
inside main() which was really clunky and just doesn't scale well.

Now, we have an Application trait that takes care of the event loop
and window creation and such and types that implement it take
care of widgets and their layout. Its still WIP but a big step
forward from where we previously where.

main() looks a lot cleaner now! ^^