gui: make the application return a view tree and implement message passing system
Instead of rendering itself on a canvas, the application now returns a widget tree which the backend logic
can interact with. This provides easier access for messages which can be created with the new
button widget. A button simply dispatches a message which can later be handled by the applcication logic.