css: Refactor Tokenizer to be Iterator<Token>
This makes logical sense because the Token::EOF variant
previously had the same meaning as None in an iterator.

It also means that we can abstract over token streams
more easily which will be necessary in a future commit
where we allow ComponentValues instead of Tokens to be
passed to the Parser.