css: Feed both tokens and componentvalues into the parser.
In CSS, the parser can operate both on Tokens and on
ComponentValues (produced by another parser or itself).

To implement this, i just merged Token into ComponentValues
(since that already contained a variant for PreservedTokens.

This is not ideal, since the ComponentValues enum is quite a bit
larger than Token and this wastes memory. But, I can't think of
a better solution right now.