dom: Create a macro that allows us to simulate inheritance
It forces structs to be #[repr(C)] and have the parent struct as their first field,
that way we can "safely" transmute pointers inbetween the two.
This does not suffice yet, more work is to be done in future commits.
(We need a build script that can parse the entire type hierarchy)