html: Cleanup the browser code and remove outdated function bodies
A lot of the parser code was written over half a year ago and is very outdated.
The spec comments are not up to par with the rest of the system and the entire dom
methods have been replaced by the actual dom inside the "dom" crate.

This commit removes any function that did not compile without the old dom.

It also implements the "insert a character" algorithm to see if the new dom is even usable.