CSS: Implement partial selector parsing
This removes a lot of code that was likely to remain unused for a long while (the stylesheet)
and comments out the parser methods that were implemented as i followed the spec but aren't actually
necessary to parse selectors. These will likely be needed in the future but in a modified form.

This commit mainly implements parts of selector parsing, namely the "complex real selector stuff".
It also overhauls the parsing api a bit.

Chunky commit but i can't be bothered to split it into smaller parts.