css: Avoid unnecessary clones in tokenizer
When we read an identifier, there is not actually a need to clone it, we can just pass a string slice back
because the css source stays in memory anyways.
The infra for this has been in place for a while (Cow<'a, str> everywhere) but we previously only used Cow::Owned.

Now, the tokenizer will emit Cow::borrowed whenever possible, which should be almost always.