url: Seperate IPv4 and IPv6 addresses into seperate types
The spec has a very loose concept of types
(everything is a host) and it makes sense to seperate
these since they are treated very differently
wrt. serialization/deserialization.

Of course, when resolving the host, they are more or less
the same but meh, that part makes up the vast
minority of the code.