graphics: yeet the render code and rethink architecture
The previous architecture was very much based on googles forma code, which is a vector renderer.
We don't need a vector renderer.
We need a generic 2d graphics renderer.
We also don't need an extremely optimized renderer, at least for now.

So now I draw more inspiration from raquote, in hopes of actually finishing this thing.
The code is much simpler.
This commit doesn't implement the actual rendering, just lays down a lot of the infrastructure
required.

The plan is to support persistent drawing state (the composition) unlike raquote which
iteratively draws stuff over the old canvas, much like the HTML Canvas api.