root: remove cssproperty-derive crate
Autogenerating css property code is fine and definitely
something that we should pursue in the future but not this way.
The approach was clunky.
For now, we write all parsing code by hand to see how it should
work and once we find common patterns, we can then go ahead and
autogenerate those parts.