web/dns: Insert additional records provided into the cache
When making a DNS request, the response might contain resolutions
that the server thinks are relevant but aren't directly related
to the query. These are sometimes used to directly provide the IP
for any nameservers named in the authoritative section for example.
To avoid redundant lookups, we now insert them into the cache directly.