base: remove parser-combinators library
This was an attempt at writing a generic parsing library
that never really worked out. For 8 months, its only user was web/http
and that wasn't really a good idea either. I still intend to build
a solid parsing library that can be used in various places but the
approach taken by parser-combinators was not the correct one.