web/url: Don't unwrap() as much
Unfortunately, a lot of the spec algorithms are clearly
written for C/C++, where a null value can just be ignored
if necessary. Rust does not allow for this, so translating the
algorithms can be a bit clunky.

Either way, with if_let_chains (which is an unstable feature) we can
mostly get away with it