web/core: Don't ever break into empty lines
Breaking into empty lines causes the line break iterator
to loop infinitely, as its never making any actual progress.

This would happen if the available screen space was very small
(so that nothing would fit) and the available text started
with a breakable character(whitespace).

To avoid this, we now trim breakable characters from the start
of the text to split.

This also fixes an issue where whitespace at the beginning of lines
would be rendered (it should not)