web/core: Fix last line in IFC sometimes overflowing
Once we had seen all available break points within a text
(and at none of them did the line exceed the available width)
we would previously just return the entire line,
disregarding the available width entirely.

This is not optimal, the line might overflow and there might
be opportunity to prevent it by using one of the earlier
break point. Only if there were no earlier break points or
the line does not overflow should we ever return the entire line
at once.