graphics/font: Don't layout compound glyphs when measuring widths
When measuring the width of some text (which is a very common operation
during layout), we don't care about the positions of individual subglyphs
or the curvature of the curves.

To avoid unnecessary work, we now have an iterator that only gives us the glyph
positions, but doesn't unwrap compound glyphs. This iterator is also internally
used by the RenderedGlyphIterator, so no code duplication is going on.