web/http: Remove pointless test
We hardcoded what a request is supposed to look like and
tested against that. This is pointless because not all valid
http requests look the same, for example, headers are case
insensitive. To accurately test for this, a more complex
setup is requred, so the test was removed for now.