util/serialization: Use unicode for json deserialization
Previously, we assumed all json input was in ASCII only,
which is obviously naive. This shouldn't come at a
significant performance loss, since we use sl-std's
ReversibleCharIterator internally (so character lookup
is still O(1))

The code also gets a whole lot cleaner because, you know,
chars are actually meant to be used this way, unlike bytes.