web/dns: Read DNS responses from a Read instance
Previously, we would always pass a byte buffer and an offset around.
(This is possible because we use udp, so data is received in one big
block instead of bit by bit.)

This had the disadvantage that the pointer needed to be kept updated at all times
and out of bound reads were common, resulting in panics.

It was not possible to use a simple Read instance because DNS domains
can contain references to other domains somewhere within the response.