web/core: Properly implement inline boxes
Previously, we could not render any kind of inline box
(and would simply panic)

Now, we can properly render them, (well decently anyways).

However, we lost the ability to select text. I think this
is reasonable, as the previous selection api was very ad-hoc
and other browser engines seem to have a smarter way
of doing it, although i do not know what that is.