base: delete taskrunner/
Taskrunner was an old attempt at an async runtime. We don't currently
use async anywhere, so it was just bitrotting for the past 10 months.

Keeping it around any longer doesnt make sense, even if we we're to
start using async, the code style has changed so much in the meantime
that it would probably make more sense to start from scratch anyways.