web/core: Move BFC logic into new BoxTree type
The old BFC implementation wasn't really in line
with the specification. Since we don't have floats
and absolutely positioned elements yet, we also only
ever constructed a single block formatting context
for the whole document, so there was some root-specific
logic inside BlockFormattingContext::new().

Instead the new box tree type acts like a BFC, but only
for the single root element. This allows the code inside
layout/flow/block.rs to be more generic and hopefully
bring it closer to the spec in the future.