web/core: Use relative coordinates for fragments
This makes it easier to move fragments inside the fragment
tree around.

One example where this is helpful is floats: Usually elements
are positioned first and then their contents are laid out.
However, floats are first laid out and then we find a place
where they fit on the page.

In theory, we only need the width to position a float but
its probably too complicated to do width->layout->height,
instead we can now do layout->width/height, because
the fragments inside the floating box can easily be moved
around.