web/core: Add basic float support
Note that the "clear" property is not yet supported.
There are also some obvious differences in rendering
compared to firefox, but those are due to a lack of
understanding of the spec on my part, I'll come back
to them later once i understand why what I'm doing
is not right.