web/core: Apply all properties to element in logical order
Previously, we would iterate the list of matched properties
in reverse order (most important first) and keep track of which
properties were set (using a hashset) to avoid less-important
values overriding more-important ones.

This wasn't perfect because shorthand properties set multiple
properties at once and its not trivial to keep track of that in
a nice way. Also we constantly need to allocate for the hashset
(no benchmarks, but it *sounds* slow)

Instead, we now iterate the list in logical order, applying
less-important rules first and then overriding them later if necessary.