crypto/tls: Refactor the way messages are sent
Previously we included abunch of io::Write
and io::Read everywhere, which forced us to deal
with io::Error - which is pointless, because 99%
of the time we need in-memory buffers anyways, so
we use io::Cursor, which cannot throw errors.

Instead, encoding and decoding now revolves around
the "Encoding" and "Decoding" traits, which provide
some syntactic sugar along with the Cursor type to
make serialization and deserialization easier.