web/core: Fix warning about unused replaced elements
Right now, replaced elements as independent formatting
contexts are never used, because we lack the implementations.
However, this does not mean they are unused - this should
not create a warning. So now, instead of doing nothing, we explicitly
discard them and then crash with a todo!(), which fixes a clippy
warning.