web: Fetch resource using ResourceLoader thread
This doesn't speed up loads by much yet, because the loads
are not scheduled as far in advance as they could be
(for example, images schedule their loads on first render, which
is wayyyyy too late) and because the ResourceLoader itself is
sequential, and therefore slow.

But it is a little improvement and it should help us unlock
more potential in the future.